+++ Leaf/Http/Headers.php
@@ -2,10 +2,9 @@
 /**
  * Leaf - a micro PHP 5 framework
  *
- * @author      Michael Darko <info@leafphp.netlify.com/#/>
+ * @author      Michael Darko <mickdd22@gmail.com>
  * @copyright   2019-2020 Michael Darko
  * @link        http://www.leafphp.netlify.com/#/
- * @license     http://www.leafphp.netlify.com/#/license
  * @version     2.6.4
  * @package     Leaf
  *