+++ Leaf/Http/Request.php
@@ -45,11 +45,22 @@ class Request
      */
     public $cookies;
 
+    /**
+     * The Request Method
+     */
+    public $requestMethod;
+
+    /**
+     * The Request Body
+     */
+    protected $request;
+
     public function __construct()
     {
         $this->requestMethod = $_SERVER['REQUEST_METHOD'];
         $handler = fopen('php://input', 'r');
         $this->request = stream_get_contents($handler);
+        $this->headers = new Headers();
     }
 
     /**
@@ -58,7 +69,7 @@ class Request
      */
     public function getMethod()
     {
-        return $this->env['REQUEST_METHOD'];
+        return $_SERVER['REQUEST_METHOD'];
     }
 
     /**
@@ -161,7 +172,7 @@ class Request
      */
     public function params($key = null, $default = null)
     {
-        $union = array_merge($this->get(), $this->post());
+        $union = $this->body();
         if ($key) {
             return isset($union[$key]) ? $union[$key] : $default;
         }
@@ -194,25 +205,24 @@ class Request
     /**
      * Get all the response data as an associative array
      */
-    public function getBody() {
+    public function body() {
         $data = json_decode($this->request, true);
 
+        $body = array();
+
         if($this->requestMethod === "GET") {
-            $body = array();
             foreach($_GET as $key => $value) {
                 $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
             }
             return count($body) > 0 ? $body : null;
         }
         if ($this->requestMethod == "POST" || $this->requestMethod == "PUT" || $this->requestMethod == "PATCH" || $this->requestMethod == "DELETE") {
-            if (isset($_POST)) {
-                $body = array();
+            if (isset($_POST) && !empty($_POST)) {
                 foreach($_POST as $key => $value) {
                     $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
                 }
                 return count($body) > 0 ? $body : null;
             } else {
-                $body = array();
                 foreach($data as $key => $value) {
                     $body[$key] = htmlspecialchars($value, ENT_QUOTES, 'UTF-8');
                 }